+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.3.3(unreleased)
 
-
+- fix: include nested error path in record field decode errors https://github.com/mununki/ppx_spice/pull/118
 
 ## 0.3.2